% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_dates}
\alias{clean_dates}
\title{Cleaning time stamps from barrier track data}
\usage{
clean_dates(ts)
}
\arguments{
\item{ts}{timestamps from barrier tracker}
}
\description{
Cleaning time stamps from barrier track data
}
