% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.crop}
\alias{.crop}
\title{Geofiltering}
\usage{
.crop(x, y)
}
\arguments{
\item{x}{the data from barrier track or atlas}

\item{y}{shapefile or bounding box}
}
\value{
filter data cropped by bounding box
}
\description{
Geofiltering
}
