% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.dataCache}
\alias{.dataCache}
\title{For caching data}
\usage{
.dataCache(cache = TRUE, inform = FALSE, reload = FALSE)
}
\arguments{
\item{cache}{caching data TRUE or FALSE}

\item{inform}{show messages of downloading data.}
}
\value{
cache or not cache data
}
\description{
For caching data
}
