% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getupdates.R
\name{get_amber_updates}
\alias{get_amber_updates}
\title{Get updates from AMBER databases in the data changes.}
\usage{
get_amber_updates(
  a = NULL,
  b = NULL,
  details = TRUE,
  colcheck,
  data.table = FALSE,
  tracker = TRUE
)
}
\arguments{
\item{details}{\code{logical} If TRUE, then a printed message on the summary
of the data changes will be provided. Otherwise the number of rows
and columns added will be provided.}

\item{colcheck}{\code{string} The where the checks needs to be viewed. Otherwise all the columns
will be checked.}

\item{data.table}{\code{logical} If TRUE, then provide a dataframe name since it will be provided.
Default is a print message of the changes.
@param tracker \code{logical} If TRUE, then barrier data will be checked. If FALSE, the
barrier atlas changes will be provided. Default \code{FALSE}}
}
\value{
Dataframe or print message of the changes in the AMBER database.
}
\description{
Get updates from AMBER databases in the data changes.
}
\examples{

\dontrun{

#print messages
get_amber_updates(tracker= TRUE)

}
}
