% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batlas.R
\name{get_barrieratlas}
\alias{get_barrieratlas}
\title{For downloading and processing barrier atlas data from AMBER database.}
\usage{
get_barrieratlas(
  type,
  country,
  depth,
  dbname,
  height,
  outlet,
  rivername,
  basinname,
  hclass,
  inform = FALSE,
  format = "data.table",
  cache = TRUE,
  bbox,
  reload = FALSE
)
}
\arguments{
\item{type}{\code{string} The type of barrier to filter out. The available options include
\code{"ramp"    "culvert" "dam"     "sluice"  "other"   "weir"    "ford"}.}

\item{country}{\code{string} Provide the country/s to be filtered. Default is all will be returned if left unfilled.}

\item{depth}{\code{string} Provide the depth/s to be filtered. Default is all will be returned if left unfilled.}

\item{dbname}{\code{string} Provide the AMBER database at country level to be filtered. Default is all will be returned if left unfilled.}

\item{height}{\code{string} Provide the height to be filtered. Default is all will be returned if left unfilled.}

\item{outlet}{\code{string} Provide the outlet to be filtered. Default is all will be returned if left unfilled.}

\item{rivername}{\code{string} Provide the river name/s to be filtered. Default is all will be returned if left unfilled.}

\item{basinname}{\code{string} Provide the river basin name/s to be filtered. Default is all will be returned if left unfilled.}

\item{hclass}{\code{string} Provide the river Hclass/es to be filtered. Default is all will be returned if left unfilled.}

\item{inform}{show messages of downloading data.}

\item{format}{\code{string} Either a data.table or coordinates converted into a point geofile
is returned. Default \code{data.table}.}

\item{cache}{caching data TRUE or FALSE}

\item{bbox}{\code{string} Either provide a bounding box with an xmin, ymin, xmax, ymax pattern. OR Provide
shapefile where the bounding box will be selected to crop the records.}
}
\value{
Return datatable or point geofile after filtering
}
\description{
For downloading and processing barrier atlas data from AMBER database.
}
\examples{

\dontrun{

  dams_austria <- get_barrieratlas(country = "AUSTRIA", type = "dam")

  #Only from Danube
  dd <- system.file("extdata", "danube.shp.zip", package = "ramber")

 db <- sf::st_read(dd, quiet=TRUE)

 d1 <- get_barrieratlas(country = "AUSTRIA", type = "dam", bbox = db)

}
}
