% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barriertracker.R
\name{get_barriertracker}
\alias{get_barriertracker}
\title{Downloading barriers AMBER database in the barrier tracker application.}
\usage{
get_barriertracker(
  type,
  country,
  riverwidth,
  fishpass,
  depth,
  width,
  height,
  subtype,
  usefulness,
  extension,
  flow,
  rivername,
  year = NULL,
  month = NULL,
  day = NULL,
  inform = FALSE,
  format = "data.table",
  cache = TRUE,
  bbox,
  reload = FALSE
)
}
\arguments{
\item{type}{\code{string} The type of barrier to filter out. The available options include
\code{"ramp"    "culvert" "dam"     "sluice"  "other"   "weir"    "ford"}.}

\item{country}{\code{string} Provide the country/s to be filtered. Default is all will be returned if left unfilled.}

\item{riverwidth}{\code{string} Provide the country/s to be filtered. Default is all will be returned if left unfilled.}

\item{fishpass}{\code{string} Provide the country/s to be filtered. Default is all will be returned if left unfilled.}

\item{depth}{\code{string} Provide the country/s to be filtered. Default is all will be returned if left unfilled.}

\item{width}{\code{string} Provide the country/s to be filtered. Default is all will be returned if left unfilled.}

\item{height}{\code{string} Provide the country/s to be filtered. Default is all will be returned if left unfilled.}

\item{subtype}{\code{string} Provide the country/s to be filtered. Default is all will be returned if left unfilled.}

\item{usefulness}{\code{string} Provide the country/s to be filtered. Default is all will be returned if left unfilled.}

\item{extension}{\code{string} Provide the country/s to be filtered. Default is all will be returned if left unfilled.}

\item{flow}{\code{string} Provide the country/s to be filtered. Default is all will be returned if left unfilled.}

\item{rivername}{\code{string} Provide the country/s to be filtered. Default is all will be returned if left unfilled.}

\item{year}{\code{string} Provide the country/s to be filtered. Default is all will be returned if left unfilled.}

\item{month}{\code{string} Provide the country/s to be filtered. Default is all will be returned if left unfilled.}

\item{day}{\code{string} Provide the country/s to be filtered. Default is all will be returned if left unfilled.}

\item{inform}{show messages of downloading data.}

\item{format}{\code{string} Either a data.table or coordinates converted into a point geofile
is returned. Default \code{data.table}.}

\item{cache}{caching data TRUE or FALSE}

\item{bbox}{\code{string} Either provide a bounding box with an xmin, ymin, xmax, ymax pattern. OR Provide
shapefile where the bounding box will be selected to crop the records.}
}
\value{
Return datatable or point geofile after filtering
}
\description{
Downloading barriers AMBER database in the barrier tracker application.
}
\examples{

\dontrun{

#dams from barrier tracker in Danube

dd <- system.file("extdata", "danube.shp.zip", package = "ramber")

db <- sf::st_read(dd, quiet=TRUE)

btat <- get_barriertracker(country = "Austria", bbox = db )

#use bounding box of danube to filter out data

ddbox <- get_barriertracker(bbox = c(xmin = 8.15250, ymin = 42.08333,
                                xmax= 29.73583,  ymax = 50.24500 ))

}


}
